this diff introduce datasink class, the class for managing the customized debug data storage pipeline.

Detials can be found in https://docs.google.com/document/d/1y_m32mKdj-OgLcLUz9TKhBW3PC3bBDYSBbeAH544EfM/edit?tab=t.0

Differential Revision: [D69583422](https://our.internmc.facebook.com/intern/diff/D69583422/)

[ghstack-poisoned]

+++ devtools/etdump/buffer_data_sink.cpp
@@ -17,7 +17,7 @@ namespace etdump {
 
 Result<size_t> BufferDataSink::write(const void* ptr, size_t length) {
   if (length == 0) {
-    return (size_t)offset_;
+    return offset_;
   }
 
   uint8_t* last_data_end = debug_buffer_.data() + offset_;
